Move all solid icons
