+++ components/atoms/icons/solid/ArrowLeftIcon.vue
