+++ components/atoms/icons/solid/ArrowNarrowRightIcon.vue
