+++ components/atoms/icons/solid/CalendarIcon.vue
