+++ components/atoms/icons/solid/CheckIcon.vue
