+++ components/atoms/icons/solid/GlobeIcon.vue
