+++ components/atoms/icons/solid/SearchIcon.vue
