+++ components/atoms/icons/solid/SwitchIcon.vue
