+++ components/atoms/icons/solid/UserIcon.vue
