+++ pages/g/_slug.vue
@@ -104,10 +104,10 @@ import getSpecName from '~/data/utils/getSpecName'
 import getClassTextColorClass from '~/data/utils/getClassTextColorClass'
 import BaseLoader from '~/components/atoms/BaseLoader.vue'
 import SpecIcon from '~/components/Ui/SpecIcon.vue'
-import ClockIcon from '~/components/icons/solid/ClockIcon.vue'
-import GlobeIcon from '~/components/icons/solid/GlobeIcon.vue'
-import CalendarIcon from '~/components/icons/solid/CalendarIcon.vue'
-import ArrowLeftIcon from '~/components/icons/solid/ArrowLeftIcon.vue'
+import ClockIcon from '~/components/atoms/icons/solid/ClockIcon.vue'
+import GlobeIcon from '~/components/atoms/icons/solid/GlobeIcon.vue'
+import CalendarIcon from '~/components/atoms/icons/solid/CalendarIcon.vue'
+import ArrowLeftIcon from '~/components/atoms/icons/solid/ArrowLeftIcon.vue'
 import guildConverter from '~/data/converters/guildConverter'
 import specializationSlug from '~/data/utils/specializationSlug'
